web/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Flask-SQLAlchemy==3.1.1
cryptography==41.0.0
argon2-cffi==23.1.0
pyotp==2.9.0
qrcode==7.4.2
gunicorn==21.2.0